# microservices-movie-app/users-service/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Swagger documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3003",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "Users related endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Users found successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "404": {
            "description": "Users not found"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "New user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User found successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update an user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users updated successfully",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Users not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "definitions": {
    "GetUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 123
        },
        "name": {
          "type": "string",
          "example": "Example Name"
        },
        "email": {
          "type": "string",
          "example": "example@gmail.com"
        }
      }
    },
    "CreateUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Example Name"
        },
        "email": {
          "type": "string",
          "example": "example@gmail.com"
        }
      }
    },
    "UpdateUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Example Name"
        },
        "email": {
          "type": "string",
          "example": "example@gmail.com"
        }
      }
    },
    "DeleteUser": {
      "type": "object",
      "properties": {}
    }
  }
}